+++ .storybook/main.cjs
@@ -8,8 +8,8 @@ module.exports = {
   "framework": "@storybook/react",
   "staticDirs": [
     {
-      from: '../theme/src/assets',
-      to: '/~@owlui/theme/src/assets'
+      from: '../lib/src/theme/assets',
+      to: '/lib/src/theme/assets'
     }
   ],
   "stories": ["../packages/**/*.stories.mdx", "../packages/**/*.stories.@(js|jsx|ts|tsx)"],
@@ -22,16 +22,7 @@ module.exports = {
       }
     }
   },
-  "addons": ['@storybook/addon-links', '@storybook/addon-essentials', {
-    name: '@storybook/addon-postcss',
-    options: {
-      postcssLoaderOptions: {
-        implementation: postcss,
-        "modules": true,
-        "plugins": postcssPlugins
-      }
-    }
-  }, '@storybook/preset-scss'],
+  "addons": ['@storybook/addon-links', '@storybook/addon-essentials', '@storybook/preset-scss'],
   // @ts-ignore
   'webpackFinal': config => {
     let ruleTest;
@@ -78,7 +69,7 @@ module.exports = {
             "../",
             "../../node_modules/",
             "../../node_modules/@owlui/theme/src/global/",
-            "../../node_modules/@owlui/theme/src/assets/",
+            "../../node_modules/@owlui/lib/src/theme/assets/",
             "../@owlui/theme/src/global/",
             "./node_modules/"
           ]