+++ CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.27](https://github.com/EEBOS/owl-ui/compare/v0.0.26...v0.0.27) (2022-07-05)
+
+**Note:** Version bump only for package @owlui/root
+
+
+
+
+
+## [0.0.26](https://github.com/EEBOS/owl-ui/compare/v0.0.25...v0.0.26) (2022-07-05)
+
+**Note:** Version bump only for package @owlui/root
+
+
+
+
+
+## [0.0.25](https://github.com/EEBOS/owl-ui/compare/v0.0.24...v0.0.25) (2022-07-04)
+
+**Note:** Version bump only for package @owlui/root
+
+
+
+
+
+## [0.0.24](https://github.com/EEBOS/owl-ui/compare/v0.0.23...v0.0.24) (2022-06-30)
+
+**Note:** Version bump only for package @owlui/root
+
+
+
+
+
 ## [0.0.23](https://github.com/EEBOS/owl-ui/compare/v0.0.22...v0.0.23) (2022-06-29)
 
 **Note:** Version bump only for package @owlui/root